Update test spec and snapshots
