--- services/minespace-web/src/tests/components/incident/IncidentsPage.spec.js
@@ -1,31 +1,0 @@
-import React from "react";
-import { shallow } from "enzyme";
-import { IncidentPage } from "@/components/pages/Incidents/IncidentPage";
-import * as MOCK from "@/tests/mocks/dataMocks";
-
-const props = {};
-const dispatchProps = {};
-
-const setupProps = () => {
-  props.match = { params: { mineGuid: "5c654fe9-bce5-4ee4-891c-806c46266d54" } };
-  props.location = {
-    state: { current: 0, mine: MOCK.MINES.mines["18133c75-49ad-4101-85f3-a43e35ae989a"] },
-  };
-};
-
-const setupDispatchProps = () => {
-  dispatchProps.clearMineIncident = jest.fn(() => Promise.resolve());
-  dispatchProps.createMineIncident = jest.fn(() => Promise.resolve());
-};
-
-beforeEach(() => {
-  setupProps();
-  setupDispatchProps();
-});
-
-describe("IncidentPage", () => {
-  it("renders properly", () => {
-    const component = shallow(<IncidentPage {...props} {...dispatchProps} />);
-    expect(component).toMatchSnapshot();
-  });
-});
